remove proxy class
