+++ can/interfaces/ixxat/__init__.py
@@ -15,7 +15,4 @@ __all__ = [
     "structures",
 ]
 
-from can.interfaces.ixxat.canlib import IXXATBus
-
-# import this and not the one from vcinpl2 for backward compatibility
-from can.interfaces.ixxat.canlib_vcinpl import get_ixxat_hwids
+from can.interfaces.ixxat.canlib import IXXATBus, get_ixxat_hwids
